Issue #26: [fix] correct type checks inside of getter and setter functions

+++ KPI_Rover/Database/ulDatabase.c
@@ -181,7 +181,7 @@ bool ulDatabase_setInt16(uint16_t id, int16_t value)
 	if (p == NULL)
 		return false;
 
-	if (p->type != UINT16)
+	if (p->type != INT16)
 		return false;
 
 	DB_LOCK();
@@ -198,7 +198,7 @@ bool ulDatabase_getInt16(uint16_t id, int16_t *value)
 	if (p == NULL)
 		return false;
 
-	if (p->type != UINT16)
+	if (p->type != INT16)
 		return false;
 
 	DB_LOCK();
@@ -215,7 +215,7 @@ bool ulDatabase_setUint32(uint16_t id, uint32_t value)
 	if (p == NULL)
 		return false;
 
-	if (p->type != UINT16)
+	if (p->type != UINT32)
 		return false;
 
 	DB_LOCK();
@@ -232,7 +232,7 @@ bool ulDatabase_getUint32(uint16_t id, uint32_t *value)
 	if (p == NULL)
 		return false;
 
-	if (p->type != UINT16)
+	if (p->type != UINT32)
 		return false;
 
 	DB_LOCK();
@@ -249,7 +249,7 @@ bool ulDatabase_setInt32(uint16_t id, int32_t value)
 	if (p == NULL)
 		return false;
 
-	if (p->type != UINT16)
+	if (p->type != INT32)
 		return false;
 
 	DB_LOCK();
@@ -266,7 +266,7 @@ bool ulDatabase_getInt32(uint16_t id, int32_t *value)
 	if (p == NULL)
 		return false;
 
-	if (p->type != UINT16)
+	if (p->type != INT32)
 		return false;
 
 	DB_LOCK();
@@ -283,7 +283,7 @@ bool ulDatabase_setFloat(uint16_t id, float value)
 	if (p == NULL)
 		return false;
 
-	if (p->type != UINT16)
+	if (p->type != FLOAT)
 		return false;
 
 	DB_LOCK();
@@ -300,7 +300,7 @@ bool ulDatabase_getFloat(uint16_t id, float *value)
 	if (p == NULL)
 		return false;
 
-	if (p->type != UINT16)
+	if (p->type != FLOAT)
 		return false;
 
 	DB_LOCK();
